Rename index.tsx to Search.tsx to match component name
